backend/backend_api/requirements.txt
django
djangorestframework
psycopg2
pillow